requirements.txt
# PyTorch + torchvision (CUDA 12.8 빌드)
torch==2.8.0
torchvision==0.23.0

# 나머지
numpy>=2,<3
transformers>=4.56,<5
accelerate>=1.11.0
open_clip_torch>=3.2.0